+++ src/app/comment/[comment_id]/edit/form.tsx
@@ -6,7 +6,7 @@ import Button from '@/components/basic/button';
 import { FinishIcon } from '@/components/icon';
 import Link from 'next/link';
 import { updateComment } from '@/app/api/comment.api';
-import type { AcdCommentEditReq } from '@/types/comment.type';
+import type { AcdCommentEditReq } from '@/types/comment.types';
 
 function Submitted({ back }: { back: string }) {
   return (
