+++ src/app/comment/[comment_id]/report/inner/form.tsx
@@ -1,5 +1,5 @@
 import Button from '@/components/basic/button';
-import type { AcdCommentReportReq } from '@/types/comment.type';
+import type { AcdCommentReportReq } from '@/types/comment.types';
 
 const reasons = [
   { reason: 'PROFANITY', text: '욕설/비방/비하' },
