+++ src/app/lecture/[id]/(root)/layout.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from 'react';
 import type { Metadata, ResolvingMetadata } from 'next';
 import { GET } from '@/app/api/get';
 import { LectureIconPath } from '@/components/composite/lectureIcon';
-import type { Course, CourseOnly } from '@/types/course.type';
+import type { Course, CourseOnly } from '@/types/course.types';
 
 interface Props {
   basic: ReactNode;
