+++ src/app/lecture/[id]/write/fetch.tsx
@@ -8,8 +8,8 @@ import { insertComment } from '@/app/api/comment.api';
 import CommentEditor from '@/components/composite/commentEditor';
 import Button from '@/components/basic/button';
 import { FinishIcon } from '@/components/icon';
-import type { AcdCommentNewReq } from '@/types/comment.type';
-import type { Course } from '@/types/course.type';
+import type { AcdCommentNewReq } from '@/types/comment.types';
+import type { Course } from '@/types/course.types';
 
 function NewCommentWithId(id: number) {
   return {
