+++ src/app/mypage/(root)/@bookmarks/components/MyBookmarksView.tsx
@@ -7,7 +7,7 @@ import CoursePreview from '@/components/view/CoursePreview';
 import { CourseLoadingItems } from '@/app/lecture/(home)/aux';
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { MyPageBookmarks } from '@/app/api/mypage.api';
-import type { Course } from '@/types/course.type';
+import type { Course } from '@/types/course.types';
 interface Props {
   totalCount: number;
 }
