+++ src/components/composite/commentEditor.tsx
@@ -6,7 +6,7 @@ import { useEffect, useRef } from 'react';
 
 import Popover from '../basic/popover';
 import { useState } from 'react';
-import { AcdCommentReqJoin } from '@/types/comment.type';
+import { AcdCommentReqJoin } from '@/types/comment.types';
 
 function FieldSetStar<Req>({
   name,
