+++ src/components/view/CommentsSummaryView/RateSummary.tsx
@@ -4,7 +4,7 @@ import Star5 from '@/components/composite/starIndicator';
 
 import { getTagFromCourse } from '@/lib/process/tag';
 import Skeleton from '@/components/composite/skeleton';
-import type { Course } from '@/types/course.type';
+import type { Course } from '@/types/course.types';
 
 function RateSummaryUnsafe(props: {
   avg_rating: React.ReactNode;
