--- src/types/InputHandler.ts
@@ -1,1 +1,0 @@
-type InputHandler = React.FormEventHandler;
