--- src/types/SelectOpts.ts
@@ -1,4 +1,0 @@
-interface SelectOpts<T> {
-  value: T;
-  label: string;
-}
