--- src/types/dto/ReqSearch.ts
@@ -1,3 +1,0 @@
-interface ReqSearch {
-  keyword: string;
-}
