+++ src/types/dto/SearchRequest.ts
@@ -1,3 +1,6 @@
+interface ReqSearch {
+  keyword: string;
+}
 interface ReqSearchCourse extends ReqSearch {
   // order: SearchOrdering;
   order: string; // TODO
