--- src/types/dto/SimpleCheckLogin.ts
@@ -1,5 +1,0 @@
-interface SimpleCheckLogin {
-  profile_id: number;
-  email: string;
-  role: string;
-}
