--- src/types/dto/UpdatePWExtended.ts
@@ -1,3 +1,0 @@
-interface UpdatePWExtended extends UpdatePWReq {
-  new_password_check: string;
-}
