--- src/types/dto/UpdatePwReq.ts
@@ -1,4 +1,0 @@
-interface UpdatePWReq {
-  old_password: string;
-  new_password: string;
-}
