+++ src/types/user.types.ts
@@ -1,0 +1,8 @@
+import z from 'zod';
+
+export const UpdatePWRequestSchema = z.object({
+  old_password: z.string(),
+  new_password: z.string(),
+});
+
+export type UpdatePWRequest = z.infer<typeof UpdatePWRequestSchema>;
